+++ src/js/common/constants.mjs
@@ -11,6 +11,9 @@ export const SETTING_TYPES = {
   devmode_tests: 'check',
   devmode_ignore_comp_checks: 'check',
   devmode_console_errors: 'check',
+  cmd_assistant_dyn_options_disabled: 'check',
+  cmd_assistant_match_mode: 'option',
+  cmd_assistant_max_results: 'int',
 };
 
 export const SETTING_NAMES = Object.keys(SETTING_TYPES);
@@ -25,6 +28,9 @@ export const SETTING_DEFAULTS = {
   devmode_tests: false,
   devmode_ignore_comp_checks: false,
   devmode_console_errors: false,
+  cmd_assistant_dyn_options_disabled: false,
+  cmd_assistant_match_mode: 'includes',
+  cmd_assistant_max_results: 35,
 };
 
 export const IGNORED_KEYS = ['Control', 'Meta', 'Shift', 'Alt', 'Escape'];