+++ src/js/page/terminal/core/command_assistant.mjs
@@ -5,8 +5,6 @@ import isEmpty from '@terminal/utils/is_empty';
 import {getArgumentInfo, getArgumentInfoByName} from '@trash/argument';
 import {ARG, INSTRUCTION_TYPE} from '@trash/constants';
 
-const MAX_OPTIONS = 35;
-
 export default class CommandAssistant {
   #virtMachine = null;
   #parent = null;
@@ -29,16 +27,14 @@ export default class CommandAssistant {
         arg_infos.push(arg_info);
       }
     }
-    return [arg_infos, command_info.args.length];
+    return arg_infos;
   }
 
   #getAvailableParameters(command_info, arg_name, arg_value) {
     const arg_info = getArgumentInfoByName(command_info.args, arg_name);
     const res_param_infos = [];
-    let total_count = 0;
     if (arg_info) {
       if (arg_info.strict_values) {
-        total_count = arg_info.strict_values.length;
         const def_value = arg_info.default_value;
         for (const strict_value of arg_info.strict_values) {
           if (!arg_value || String(strict_value).startsWith(arg_value)) {
@@ -56,7 +52,6 @@ export default class CommandAssistant {
         arg_info.default_value &&
         String(arg_info.default_value).startsWith(arg_value)
       ) {
-        total_count = 1;
         res_param_infos.push({
           value: arg_info.default_value,
           is_default: true,
@@ -68,15 +63,19 @@ export default class CommandAssistant {
       }
     }
 
-    return [res_param_infos, total_count];
+    return res_param_infos;
   }
 
   #filterParameterOptions(values, filter_by) {
     let res = values || [];
     if (filter_by) {
-      res = values.filter(item => item.startsWith(filter_by));
+      if (this.#parent.config.cmd_assistant_match_mode === 'includes') {
+        res = values.filter(item => item.includes(filter_by));
+      } else {
+        res = values.filter(item => item.startsWith(filter_by));
+      }
     }
-    return res.slice(0, MAX_OPTIONS);
+    return res;
   }
 
   async #getAvailableDynamicParameters(command_info, arg_name, arg_value) {
@@ -99,7 +98,7 @@ export default class CommandAssistant {
         values: options,
       });
     }
-    return [ret, options.length];
+    return ret;
   }
 
   getSelectedParameterIndex(parse_info, caret_pos) {
@@ -194,12 +193,12 @@ export default class CommandAssistant {
   }
 
   async getAvailableOptions(data, caret_pos) {
+    const ret = [];
     const input_info = this.getInputInfo(data, caret_pos);
     if (isEmpty(input_info)) {
-      return [[], 0];
+      return ret;
     }
     if (input_info.index.cmd === input_info.index.current) {
-      const ret = [];
       // Command name
       const cmd_names = this.#getAvailableCommandNames(
         input_info.token.cmd || data,
@@ -212,20 +211,19 @@ export default class CommandAssistant {
           description: this.#virtMachine.commands[cmd_name].definition,
         });
       }
-      return [ret, Object.keys(this.#virtMachine.commands).length];
+      return ret;
     }
 
     const command_info = input_info.token.cmd
       ? this.#virtMachine.commands[input_info.token.cmd]
       : undefined;
     if (!command_info) {
-      return [[], 0];
+      return [];
     }
 
     if (input_info.index.current === input_info.index.arg) {
-      const ret = [];
       // Argument
-      const [arg_infos, total_count] = this.#getAvailableArguments(
+      const arg_infos = this.#getAvailableArguments(
         command_info,
         input_info.token.arg,
       );
@@ -240,21 +238,22 @@ export default class CommandAssistant {
           values: arg_info.strict_values,
         });
       }
-      return [ret, total_count];
     } else if (
       input_info.index.current !== input_info.index.cmd &&
       input_info.index.current - 1 === input_info.index.arg &&
       input_info.token.arg
     ) {
-      const ret = [];
       // Parameter
-      let [param_infos, total_count] = this.#getAvailableParameters(
+      let param_infos = this.#getAvailableParameters(
         command_info,
         input_info.token.arg,
         input_info.token.current,
       );
-      if (isEmpty(param_infos)) {
-        [param_infos, total_count] = await this.#getAvailableDynamicParameters(
+      if (
+        !this.#parent.config.cmd_assistant_dyn_options_disabled &&
+        isEmpty(param_infos)
+      ) {
+        param_infos = await this.#getAvailableDynamicParameters(
           command_info,
           input_info.token.arg,
           input_info.token.current,
@@ -272,9 +271,8 @@ export default class CommandAssistant {
           values: param_info.strict_values,
         });
       }
-      return [ret, total_count];
     }
 
-    return [[], 0];
+    return ret;
   }
 }